refractor: 일기 발송 기록 id 타입 Long으로 변환

+++ ondam-backend/src/main/java/com/piveguyz/ondambackend/diaryRecord/command/application/dto/DiaryRecordDTO.java
@@ -4,9 +4,9 @@ import lombok.Data;
 
 @Data
 public class DiaryRecordDTO {
-    private Integer id;
-    private Integer diaryId;
-    private Integer senderId;
-    private Integer receiverId;
+    private Long id;
+    private Long diaryId;
+    private Long senderId;
+    private Long receiverId;
     private String isExpired;
 }
